+++ skfem/element/element_tet/__init__.py
@@ -2,7 +2,7 @@ from .element_tet_p0 import ElementTetP0  # noqa
 from .element_tet_p1 import ElementTetP1  # noqa
 from .element_tet_p2 import ElementTetP2  # noqa
 from .element_tet_rt1 import ElementTetRT1  # noqa
-from .element_tet_n0 import ElementTetN0  # noqa
+from .element_tet_n1 import ElementTetN1  # noqa
 from .element_tet_mini import ElementTetMini  # noqa
 from .element_tet_cr import ElementTetCR  # noqa
 from .element_tet_ccr import ElementTetCCR  # noqa
