+++ tests/test_assembly.py
@@ -16,8 +16,8 @@ from skfem.element import (ElementQuad1, ElementQuadS2, ElementHex1,
                            ElementTetDG, ElementTriHermite, ElementVector,
                            ElementTriRT1, ElementTriRT2, ElementTriBDM1,
                            ElementQuadRT1, ElementTetRT1, ElementHexRT1,
-                           ElementTriN0, ElementTriP0, ElementTetN0,
-                           ElementQuadN0, ElementQuad0, ElementTriN1)
+                           ElementTriN1, ElementTriP0, ElementTetN0,
+                           ElementQuadN1, ElementQuad0, ElementTriN2)
 from skfem.mesh import (MeshQuad, MeshHex, MeshTet, MeshTri, MeshQuad2,
                         MeshTri2, MeshTet2, MeshHex2, MeshTri1DG, MeshQuad1DG,
                         MeshHex1DG)
@@ -654,9 +654,9 @@ def test_hdiv_boundary_integration_3d(basis):
 @pytest.mark.parametrize(
     "mtype,e1,e2,e3",
     [
-        (MeshTri, ElementVector(ElementTriP1()), ElementTriN0(), ElementTriP0()),
-        (MeshTri, ElementVector(ElementTriP2()), ElementTriN1(), ElementTriP0()),
-        (MeshQuad, ElementVector(ElementQuad1()), ElementQuadN0(), ElementQuad0()),
+        (MeshTri, ElementVector(ElementTriP1()), ElementTriN1(), ElementTriP0()),
+        (MeshTri, ElementVector(ElementTriP2()), ElementTriN2(), ElementTriP0()),
+        (MeshQuad, ElementVector(ElementQuad1()), ElementQuadN1(), ElementQuad0()),
     ]
 )
 def test_hcurl_projections_2d(mtype, e1, e2, e3):
